many improvements and features, among which the most notable is the language model based on Efron's work. It works properly at the moment, just does not make much sense.
Rest gathering works flawlessly except it cannot be stopped until it finishes gathering.

naive expansion works, efron's method is on its way.

program still can easily crash from being overcrowded with tweets, so limits should be placed on the UI but not on the database gathering, that works flawlessly.

keywords should also update in a separate thread so that there is no lag in receiving tweets.